Move types related to source files to submodule
