+++ experiments/rust-minimal/src/lib.rs
@@ -3,6 +3,7 @@
 // Supporting modules
 mod alloc;
 pub mod env;
+pub mod source;
 
 // Intermediate languages
 pub mod core;
@@ -22,36 +23,3 @@ pub type StringInterner = string_interner::StringInterner<
     string_interner::backend::BucketBackend<StringId>,
     std::hash::BuildHasherDefault<fxhash::FxHasher32>,
 >;
-
-/// File id.
-pub type FileId = usize;
-
-/// Byte offsets into source files.
-pub type BytePos = usize;
-
-/// Byte ranges in source files.
-#[derive(Debug, Copy, Clone)]
-pub struct ByteRange {
-    start: BytePos,
-    end: BytePos,
-}
-
-impl ByteRange {
-    pub const fn new(start: BytePos, end: BytePos) -> ByteRange {
-        ByteRange { start, end }
-    }
-
-    pub const fn start(&self) -> BytePos {
-        self.start
-    }
-
-    pub const fn end(&self) -> BytePos {
-        self.end
-    }
-}
-
-impl Into<std::ops::Range<usize>> for ByteRange {
-    fn into(self) -> std::ops::Range<usize> {
-        self.start..self.end
-    }
-}
